[#834] Dropped a redundant cast around the pinned clock offset.
PHPStan narrows 'strtotime()' to an integer when its offset is a literal, so the cast said nothing the surrounding data provider does not already rely on.

+++ tests/phpunit/src/Unit/Steps/Generic/DateTraitTest.php
@@ -121,7 +121,7 @@ class DateTraitTest extends UnitTestCase {
   public function testUnskippedScenarioResolvesTokens(): void {
     $this->testObject->dateBeforeScenario($this->createBeforeScenarioScope());
 
-    $expected = date('Y-m-d', (int) strtotime('-1 day', DateTraitTestImplementation::CLOCK));
+    $expected = date('Y-m-d', strtotime('-1 day', DateTraitTestImplementation::CLOCK));
 
     $this->assertSame($expected, $this->testObject->dateRelativeTransformValue('[relative:-1 day#Y-m-d]'));
